Allow to override name of db:migrate rake task for Rails.
This setting might be useful when other solutions based on schema.rb
are used for database migrations. Especially when someone uses a gem
like apartment to deal with it and it provides a different rake task.

In case of apartment there is a rake task called apartment:migrate.

My idea to make it a setting was to use the built-in check of changes
in db/schema.rb rather than implement it in a very similar way one
more time.